Greencoin signer/verifier v0
The first version of a Greencoin signer/verifier for offline use of transaction generation and public/private key generation. This application is currently only compiled for windows but will be compiled for Linux soon. This application creates keys/ID's in the Greencoin format. Greencoin recommends all private keys be kept offline and this application does not save your private keys anywhere.